+++ src/main/java/com/PeaceOfMind/MicroMarket/entity/DetalleVenta.java
@@ -1,5 +1,10 @@
 package com.PeaceOfMind.MicroMarket.entity;
 
+import java.math.BigDecimal;
+
+import org.hibernate.annotations.SQLDelete;
+import org.hibernate.annotations.SQLRestriction;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -9,9 +14,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
-
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.SQLRestriction;
 
 @SQLDelete(sql = "UPDATE Detalle_Venta SET deleted = true WHERE id_detalle = ?")
 @SQLRestriction("deleted = false")
@@ -38,10 +40,10 @@ public class DetalleVenta {
     private Integer cantidad;
 
     @Column(name = "precio_unitario", nullable = false, precision = 10, scale = 2)
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private Double subtotal;
+    private BigDecimal subtotal;
 
     @Column(nullable = false)
     private Boolean deleted = false;
