+++ src/main/java/com/PeaceOfMind/MicroMarket/entity/Empleado.java
@@ -1,11 +1,20 @@
 package com.PeaceOfMind.MicroMarket.entity;
 
-import jakarta.persistence.*;
-import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLRestriction;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Data;
 
 @SQLDelete(sql = "UPDATE Empleados SET deleted = true WHERE id_empleado = ?")
 @SQLRestriction("deleted = false")
@@ -34,7 +43,7 @@ public class Empleado {
     private LocalDate fechaIngreso;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private Double salario;
+    private BigDecimal salario;
 
     @Column(nullable = false)
     private Boolean deleted = false;
