+++ src/main/java/com/PeaceOfMind/MicroMarket/entity/Venta.java
@@ -1,13 +1,24 @@
 package com.PeaceOfMind.MicroMarket.entity;
 
-
-import jakarta.persistence.*;
-import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLRestriction;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.Data;
 
 @SQLDelete(sql = "UPDATE Ventas SET deleted = true WHERE id_venta = ?")
 @SQLRestriction("deleted = false")
@@ -30,13 +41,13 @@ public class Venta {
     private Empleado empleado;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private Double subtotal;
+    private BigDecimal subtotal;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private Double iva;
+    private BigDecimal iva;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private Double total;
+    private BigDecimal total;
 
     @OneToMany(mappedBy = "venta", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<DetalleVenta> detalles;
